Complete service Setters for id and routeId in WayPoint Builder was changed to long instead of int.

+++ database/routes/src/main/java/ua/vertex/waypoint/Entity/WayPoint.java
@@ -28,11 +28,11 @@ public class WayPoint implements Serializable, Comparable<WayPoint> {
         private Builder() {
         }
 
-        public Builder setId(int id) {
+        public Builder setId(long id) {
             WayPoint.this.id = id;
             return this;
         }
-        public Builder setRouteId(int routeId) {
+        public Builder setRouteId(long routeId) {
             WayPoint.this.routeId = routeId;
             return this;
         }
